+++ src/main/java/edu/scau/pyx/ssp/entity/SensitiveKeyword.java
@@ -1,22 +1,22 @@
 package edu.scau.pyx.ssp.entity;
 
 public class SensitiveKeyword {
-    private int id;
+    private long id;
     private String keyword;
 
     public SensitiveKeyword() {
     }
 
-    public SensitiveKeyword(int id, String keyword) {
+    public SensitiveKeyword(long id, String keyword) {
         this.id = id;
         this.keyword = keyword;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
